# Conflicts:
#	ElaWidgetTools/ElaSuggestBox.cpp
#	ElaWidgetTools/include/ElaSuggestBox.h
#	ElaWidgetTools/private/ElaTabWidgetPrivate.cpp

+++ ElaWidgetTools/DeveloperComponents/ElaSuggestDelegate.cpp
@@ -1,16 +1,18 @@
 #include "ElaSuggestDelegate.h"
 
-#include <QPainter>
-#include <QPainterPath>
-
+#include "ElaSuggestBox.h"
 #include "ElaSuggestBoxPrivate.h"
 #include "ElaSuggestModel.h"
 #include "ElaTheme.h"
+#include <QPainter>
+#include <QPainterPath>
 ElaSuggestDelegate::ElaSuggestDelegate(QObject* parent)
     : QStyledItemDelegate{parent}
 {
     _themeMode = eTheme->getThemeMode();
-    connect(eTheme, &ElaTheme::themeModeChanged, this, [=](ElaThemeType::ThemeMode themeMode) { _themeMode = themeMode; });
+    connect(eTheme, &ElaTheme::themeModeChanged, this, [=](ElaThemeType::ThemeMode themeMode) {
+        _themeMode = themeMode;
+    });
 }
 
 ElaSuggestDelegate::~ElaSuggestDelegate()
@@ -66,7 +68,7 @@ void ElaSuggestDelegate::paint(QPainter* painter, const QStyleOptionViewItem& op
         QFont iconFont = QFont("ElaAwesome");
         iconFont.setPixelSize(17);
         painter->setFont(iconFont);
-        painter->drawText(option.rect.x() + 11, option.rect.y() + 26, QChar((unsigned short)suggest->getElaIcon()));
+        painter->drawText(option.rect.x() + 11, option.rect.y() + 26, QChar(suggest->getElaIcon()));
     }
     painter->restore();
 }
